# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "nube api docs",
        "title": "GO Nube API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
        "/auth/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CheckToken header example",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "CheckToken",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "RefreshToken Process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "parameters": [
                    {
                        "description": "RefreshToken",
                        "name": "RefreshToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Signin Process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signin",
                "parameters": [
                    {
                        "description": "Signin",
                        "name": "Signin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/devices/": {
            "get": {
                "description": "get devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List devices",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "/devices/?Limit=2",
                        "name": "Limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "/devices/?Limit=0",
                        "name": "Offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "/devices/?Sort=1",
                        "name": "Sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "/devices/?Order=DESC",
                        "name": "Order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "/devices/?Search=",
                        "name": "Search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DeviceBody"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Create Device",
                "parameters": [
                    {
                        "description": "input body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeviceBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DeviceBody"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "get by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Show a Device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DeviceBody"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{uuid}": {
            "delete": {
                "description": "Delete Device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Delete Device",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "patch": {
                "description": "Update Device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Update Device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device",
                        "name": "Network",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/networks/{uuid}": {
            "get": {
                "description": "get by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Show a Network",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Network"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update Network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Update Network",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Network ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Network",
                        "name": "Network",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "description": "Delete Network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Delete Network",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Network ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Device": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "network_uuid": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Point"
                    }
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.DeviceBody": {
            "type": "object",
            "required": [
                "NetworkID",
                "name"
            ],
            "properties": {
                "NetworkID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Network": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Device"
                    }
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.Point": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device_uuid": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "point_store": {
                    "$ref": "#/definitions/model.PointStore"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "writeable": {
                    "type": "boolean"
                }
            }
        },
        "model.PointStore": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "pointUuid": {
                    "type": "string"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}